refactor(Polls): rename and split off classes into files
